+++ src/components/InstancesList/InstancesList.js
@@ -28,7 +28,10 @@ import {
   checkIfUserInCentralTenant,
   TitleManager,
 } from '@folio/stripes/core';
-import { SearchAndSort } from '@folio/stripes/smart-components';
+import {
+  SearchAndSort,
+  buildUrl,
+} from '@folio/stripes/smart-components';
 import {
   Button,
   Icon,
@@ -634,6 +637,12 @@ class InstancesList extends React.Component {
     });
   }
 
+  focusSearchField = () => {
+    setTimeout(() => {
+      this.inputRef.current?.focus();
+    });
+  }
+
   handleFastAddModalClose = ({ instanceRecord } = {}) => {
     const {
       history,
@@ -645,6 +654,8 @@ class InstancesList extends React.Component {
         pathname: `/inventory/view/${instanceRecord.id}`,
         search: location.search,
       });
+    } else {
+      this.focusSearchField();
     }
 
     this.toggleNewFastAddModal();
@@ -1133,6 +1144,32 @@ class InstancesList extends React.Component {
     return this.findAndOpenItem(instance);
   }
 
+  handleCloseNewRecord = (e) => {
+    const {
+      location,
+      match: {
+        path,
+      },
+      history,
+    } = this.props;
+
+    e?.preventDefault();
+
+    const url = buildUrl(location, { layer: null }, path);
+
+    history.push(url);
+    this.focusSearchField();
+  }
+
+  handleDismissDetail = (resetSelectedItem) => {
+    const { location } = this.props;
+    const id = this.getInstanceIdFromLocation(location);
+
+    resetSelectedItem();
+    // focus on the title of the closed record in the results list
+    document.getElementById(`record-title-${id}`)?.focus();
+  }
+
   render() {
     const {
       canUseSingleRecordImport,
@@ -1192,6 +1229,7 @@ class InstancesList extends React.Component {
               iconAlignment="baseline"
             >
               <TextLink
+                id={`record-title-${id}`}
                 to={this.getRowURL(id)}
               >
                 {title}
@@ -1284,6 +1322,7 @@ class InstancesList extends React.Component {
             inputType="textarea"
             inputRef={this.inputRef}
             indexRef={this.indexRef}
+            isCursorAtEnd
             resultCountIncrement={RESULT_COUNT_INCREMENT}
             viewRecordComponent={ViewInstanceWrapper}
             editRecordComponent={InstanceForm}
@@ -1343,6 +1382,8 @@ class InstancesList extends React.Component {
             resultsOnMarkPosition={this.onMarkPosition}
             resultsOnResetMarkedPosition={this.resetMarkedPosition}
             resultsCachedPosition={itemToView}
+            onCloseNewRecord={this.handleCloseNewRecord}
+            onDismissDetail={this.handleDismissDetail}
           />
         </div>
         <ErrorModal